dsdsd

+++ api/app.py
@@ -22,7 +22,7 @@ if tokenizer.pad_token is None:
     tokenizer.pad_token = tokenizer.eos_token
 model = AutoModelForCausalLM.from_pretrained(
     MODEL_NAME,
-    torch_dtype=torch.float16,  # Usa float16 si la GPU lo permite
+    torch_dtype=torch.bfloat16,  # Usa float16 si la GPU lo permite
     device_map="auto"
 )
